move scalar tests
